style: rename nuclear-api
